requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.5.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
openai==0.28.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pdf-redactor==0.0.1
pdfrw==0.4
pillow==11.2.1
propcache==0.3.2
prov==2.0.2
puremagic==1.29
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
PyMuPDF==1.26.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
scipy==1.15.3
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
traits==7.0.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1